refactor : Board class에서 move 역할을 새로운 클래스로 분리
