+++ README.md
@@ -119,7 +119,7 @@
 
 ## 리펙토링 목록
 
-- [ ]  Board class에서 move를 새로운 클래스로 분리
+- [x]  Board class에서 move를 새로운 클래스로 분리
 - [x]  인터페이스 Piece를 없애고 추상 클래스인 FullPiece를 Piece로 대체하여 사용하기
 - [x]  position에 `toRankValue` 를 랭크 안에서 하도록 수정
 - [ ]  매직 넘버 상수화
